Updated android project to address Firebase API changes.
